I added the different files that I had been working with to mess around. I am working on middle line now. I am having issues because when I print the image the coordinates on the image are different than the coordinates in python
In the image, it is more than 1000x1000 pixels, so there is some scaling that cv2 does. When I print the middle of the line, that the drone would ideally use, it prints it without the proper horizontal/vertical scaling so it is not in the centre of the largest line even though the coordiantes are correct.

I will keep messing around with this tomorrow morning. I think that I can use some internet sleuthing and figure this out.

Mac